+++ src/fr/ign/georeden/algorithms/string/IStringComparison.java
@@ -1,5 +1,5 @@
 package fr.ign.georeden.algorithms.string;
 
 public interface IStringComparison {
-	public float computeSimilarity(String s1, String s2);
+	public double computeSimilarity(String s1, String s2);
 }
