+++ src/fr/ign/georeden/algorithms/string/StringComparisonDamLev.java
@@ -22,17 +22,17 @@ public class StringComparisonDamLev implements IStringComparison {
 	 * @param s2
 	 * @return
 	 */
-	public float lambdaIJ(String s1, String s2) {
-		return 1 - ((float)damLev.execute(s1, s2) / Math.max(s1.length(), s2.length()));
+	public double lambdaIJ(String s1, String s2) {
+		return 1 - ((double)damLev.execute(s1, s2) / Math.max(s1.length(), s2.length()));
 	}
 	
-	public float[][] computeMatrix(String[] tokenArray1, String[] tokenArray2) {
-		float[][] matrix = new float[tokenArray1.length][tokenArray2.length];
+	public double[][] computeMatrix(String[] tokenArray1, String[] tokenArray2) {
+		double[][] matrix = new double[tokenArray1.length][tokenArray2.length];
 		for (int i = 0; i < tokenArray1.length; i++) {
 			String token1 = tokenArray1[i];
 			for (int j = 0; j < tokenArray2.length; j++) {
 				String token2 = tokenArray2[j];
-				float score = Math.round(lambdaIJ(token1, token2) * 10000f) / 10000f;
+				double score = Math.round(lambdaIJ(token1, token2) * 10000f) / 10000f;
 				matrix[i][j] = score;
 			}
 		}
@@ -40,16 +40,16 @@ public class StringComparisonDamLev implements IStringComparison {
 	}
 
 	@Override	
-	public float computeSimilarity(String s1, String s2) {
+	public double computeSimilarity(String s1, String s2) {
 		if (s1 == null || s2 == null)
-			return 0;
+			return 0.0;
 		String[] tokenArray1 = AlgoUtils.tokenizeString(s1);
 		String[] tokenArray2 = AlgoUtils.tokenizeString(s2);
-		float[][] matrix = computeMatrix(tokenArray1, tokenArray2);
-		float mu = (tokenArray1.length +tokenArray2.length) / 2f;
+		double[][] matrix = computeMatrix(tokenArray1, tokenArray2);
+		double mu = (tokenArray1.length +tokenArray2.length) / 2f;
 		HashMap<Integer, Integer> optimums = new HashMap<>();
 		for (int j = 0; j < matrix[0].length; j++) {
-			float columnMax = 0;
+			double columnMax = 0;
 			int iValue = -1;
 			for (int i = 0; i < matrix.length; i++) {
 				if (!optimums.containsValue(i) && matrix[i][j] > columnMax) {
@@ -61,7 +61,7 @@ public class StringComparisonDamLev implements IStringComparison {
 				optimums.put(j, iValue);
 		}
 		
-		float sum = 0f;
+		double sum = 0f;
 		for (Iterator<Integer> iterator = optimums.keySet().iterator(); iterator.hasNext();) {
 			Integer j = iterator.next();
 			Integer i = optimums.get(j);
@@ -72,7 +72,7 @@ public class StringComparisonDamLev implements IStringComparison {
 		return sum / mu;
 	}
 	
-	public void displayMatrix(float[][] matrix) {
+	public void displayMatrix(double[][] matrix) {
 		String line = "";
 		for (int j = 0; j < matrix[0].length; j++) {
 			for (int i = 0; i < matrix.length; i++) {
